# storage: improve `ComputeStats` performance